scout-dbt/requirements.txt
dbt-core>=1.7.0,<1.9.0
dbt-postgres>=1.7.0,<1.9.0
dbt-sqlserver>=1.8.0,<1.9.0
pandas==2.1.0
pyarrow==14.0.0
psycopg2-binary==2.9.9
pyodbc==5.0.0
azure-storage-blob==12.19.0
